refactor: move ScopeQueryParam to separate file
