+++ projects/observability/src/pages/explorer/explorer.types.ts
@@ -1,0 +1,4 @@
+export const enum ScopeQueryParam {
+  EndpointTraces = 'endpoint-traces',
+  Spans = 'spans'
+}
